Move addArgCond for predator page into resolver class
